fix: type conversion, should be float64 instead of int

+++ routes/base.go
@@ -138,7 +138,7 @@ func Initialize(cfg config.Config, ctx *types.Context) (*fiber.App, error) {
 
 		user := c.Locals("user").(*jwt.Token)
 		claims := user.Claims.(jwt.MapClaims)
-		userId := claims["id"].(int)
+		userId := claims["id"].(float64)
 		username := claims["user"].(string)
 		ctx.Database.Create(&types.SpotifyAuthToken{Id: userId, Username: username, Token: spotifyToken.Token})
 		return c.SendStatus(fiber.StatusAccepted)
